move chips to separate file, wrap chips when too big or many, change text to black
